Update isomorphic-git
Interested in getting https://github.com/isomorphic-git/isomorphic-git/pull/1848 which fixes https://github.com/isomorphic-git/isomorphic-git/issues/1847.

This item from the notes in 1847 particularly catches my eye. "If an error occurs before the PACK record header, isomorphic-git will log it to STDOUT, but proceed without a packfile, which may lead to the previous error or an error when reading the repository. (GitPackIndex.fromPack)"